% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anestheticdat.R
\docType{data}
\name{Anestheticdat}
\alias{Anestheticdat}
\title{Data for the Anesthetic example in Greco et al. (2013)}
\format{An object of class \code{data.frame} with 30 rows and 10 columns.}
\source{
Greco, T., Landoni, G., Biondi-Zoccai, G., D'Ascenzo, F. and Zangrillo, A., 2013.
A Bayesian network meta-analysis for binary outcome: how to do it. Statistical methods in
medical research, p.0962280213500185.
}
\usage{
Anestheticdat
}
\description{
Greco et al. (2013) considered four anaesthetic agents used in cardiac surgery
(called here 1, 2, 3 and 4) and investigated through NMA.
}
\keyword{datasets}
