% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Certolizumabdat.R
\docType{data}
\name{Certolizumabdat}
\alias{Certolizumabdat}
\title{Data for the Certolizumab NMA-network discussed in Dias et al. (2013)}
\format{
An object of class \code{data.frame} with 12 rows and 9 columns.
}
\source{
Dias, S., Sutton, A.J., Welton, N.J. and Ades, A.E., 2013.
Evidence synthesis for decision making 3 heterogeneity-subgroups,
meta-regression, bias, and bias-adjustment. Medical Decision Making, 33(5),
pp.618-640.
}
\usage{
Certolizumabdat
}
\description{
Data are available from a review of trials of certolizumab pegol (CZP)
for the treatment of rheumatoid arthritis in patients who had failed
on disease-modifying antirheumatic drugs, including methotrexate (MTX). 20
Twelve MTX controlled trials were identified, comparing 6 different
treatments with placebo. The primary outcome is improving by at least 50% on the American College
of Rheumatology scale (ACR50) at 6 months. A trial-specificc covariate, the mean disease duration
in years for patients, is also given.
}
\keyword{datasets}
