% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diabetesdat.R
\docType{data}
\name{Diabetesdat}
\alias{Diabetesdat}
\title{Data for the Diabetes example in Senn et al. (2013)}
\format{An object of class \code{data.frame} with 26 rows and 11 columns.}
\source{
Senn, S., Gavini, F., Magrez, D., & Scheen, A. (2013). Issues in
performing a network meta-analysis. Statistical Methods in Medical
Research, 22, 169--189.
}
\usage{
Diabetesdat
}
\description{
Results from 26 trials examining the effectiveness of glucose-lowering agents in
patients with type 2 diabetes.
}
\keyword{datasets}
