% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dietaryfatdat.R
\docType{data}
\name{Dietaryfatdat}
\alias{Dietaryfatdat}
\title{Data for the Dietary fat example in Dias et al. (2011)}
\format{An object of class \code{data.frame} with 10 rows and 10 columns.}
\source{
Dias, S., Welton, N.J., Sutton, A.J. & Ades, A.E. NICE DSU Technical Support Document 2:
A Generalised Linear Modelling Framework for Pairwise and Network Meta-Analysis of
Randomised Controlled Trials. 2011; last updated September 2016; available from
http://www.nicedsu.org.uk
}
\usage{
Dietaryfatdat
}
\description{
In a Cochrane Review of randomised controlled trials to assess the effect of change in dietary
fats on total and cardiovascular mortality, 104 data extracted was in the form of rates and given
as the number of events per person-years observed.
}
\keyword{datasets}

