% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncDiabetesdat.R
\docType{data}
\name{IncDiabetesdat}
\alias{IncDiabetesdat}
\title{Data for the Incident Diabetes example in Elliott et al. (2007)}
\format{
An object of class \code{data.frame} with 22 rows and 12 columns.
}
\source{
Elliott WJ and Meyer PM (2007). "Incident diabetes in clinical trials of antihypertensive
 drugs: a network meta-analysis." Lancet 369(9557), 201-7.
}
\usage{
IncDiabetesdat
}
\description{
An example of network meta-analysis for binary outcomes with follow-up times reported.
}
\details{
This network meta-analysis is studied by Elliott and Meyer (2007) to assess the effects of an
ti-hypertensive agents on incident diabetes. Treatment IDs represent 1) diuretic; 2) placebo;
3) beta-blocker; 4) CCB; 5) ACE inhibitor; and 6) ARB.
}
\keyword{datasets}
