% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KussHeartdat.R
\docType{data}
\name{KussHeartdat}
\alias{KussHeartdat}
\title{Data for the ischemic heart disease sparse pairwise meta-analysis discussed in Kuss (2014)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 60 rows and 4 columns.
}
\source{
Kuss O. (2015), Statistical methods for meta-analyses including information
from studies without any events--add nothing to nothing and succeed nevertheless,
Statist. Med., 34; pages 1097-1116, doi: 10.1002/sim.6383
}
\usage{
KussHeartdat
}
\description{
Dataset for pairwise meta-analysis with many zero entries. Data is from a Cochrane review on
postoperative stroke occurrence when comparing off-pump and on-pump coronary artery
bypass grafting for ischemic heart disease. This data is used as an example in Kuss (2014).
}
\keyword{datasets}
