% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parkinsondat.R
\docType{data}
\name{Parkinsondat}
\alias{Parkinsondat}
\title{Data for the Parkinson NMA-network discussed in Dias et al. (2013)}
\format{An object of class \code{data.frame} with 7 rows and 11 columns.}
\source{
Dias, S., Welton, N.J., Sutton, A.J. & Ades, A.E. NICE DSU Technical Support Document 2:
A Generalised Linear Modelling Framework for Pairwise and Network Meta-Analysis of
Randomised Controlled Trials. 2011; last updated September 2016; available from
http://www.nicedsu.org.uk
}
\usage{
Parkinsondat
}
\description{
Data are the mean off-time reduction in patients given dopamine
Agonists as adjunct therapy in Parkinson's disease.
The data available are the mean, standard deviation and number of patients
in each trial arm, for 7 studies of five different drugs: placebo,
coded 1, and five active drugs coded 2 to 5.
}
\keyword{datasets}

