% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nma.R
\name{plot_nma}
\alias{plot_nma}
\title{Plot a network plot}
\source{
This function is taken from \code{nma.networkplot} function from \code{pcnetmeta} R package.
}
\usage{
plot_nma(s.id, t.id, data, title = "", adjust.figsizex = 1.1,
  adjust.figsizey = 1.1)
}
\arguments{
\item{s.id}{Variable holding the study IDs for each study.}

\item{t.id}{Variable holding the treatments for each study.}

\item{data}{A \code{create_INLA_dat} object.}

\item{title}{A character string indicating plot title.}

\item{adjust.figsizex}{a positive number used to adjust the plot width. The default is 1.1.}

\item{adjust.figsizey}{a positive number used to adjust the plot height. The default is 1.1.}
}
\description{
Takes a \code{create_INLA_dat} output and plots a network graph.
}
\seealso{
\code{pcnetmeta::nma.networkplot}
}
\author{
Lifeng Lin, Jing Zhang, and Haitao Chu
}
