% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.nma_inla}
\alias{print.nma_inla}
\title{Print nmainla object}
\usage{
\method{print}{nma_inla}(x, digits = 3, ...)
}
\arguments{
\item{x}{An \code{nma_inla} object.}

\item{digits}{An integer indicating the number of decimal places.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The return value is invisible \code{NULL}
}
\description{
Takes an \code{nma_inla} object which is obtained by function \code{nma_inla} and print
the model and data information such as model type used in the model.
}
\details{
The resulting data.frame can be used as data argument in \code{nma_inla}.
}
