% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strokedat.R
\docType{data}
\name{Strokedat}
\alias{Strokedat}
\title{Data for the Stroke NMA regression discussed in Batson et al. (2016)}
\format{
An object of class \code{data.frame} with 19 rows and 18 columns.
}
\source{
Batson S, Sutton A, Abrams K (2016) Exploratory Network Meta Regression Analysis of
Stroke Prevention in Atrial Fibrillation Fails to Identify
Any Interactions with Treatment Effect. PLoS ONE 11
}
\usage{
Strokedat
}
\description{
Dataset for network meta-regression of stroke prevention in Atrial Fibrillation.
A total of 19 studies, and primary endpoint is reported ischaemic stroke. It includes 15 comparators
which include fixed low dose warfarin with or without aspirin, aspirin monotherapy, aspirin plus
clopidogrel, indobufen, idraparinux, triflusal and ximelagatran. Study level covariates are the proportion of
patients with a previous stroke/TIA, proportion of males, mean age, and the duration of study.
}
\keyword{datasets}
