% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Woodsdat.R
\docType{data}
\name{Woodsdat}
\alias{Woodsdat}
\title{Data for the Woods example in Woods et al. (2010)}
\format{An object of class \code{data.frame} with 3 rows and 14 columns.}
\source{
Woods BS, Hawkins N, Scott DA (2010). Network meta-analysis on the log-hazard scale, 
combining count and hazard ratio statistics accounting for multi-arm trials: A tutorial. 
BMC Medical Research Methodology 10, 54.
}
\usage{
Woodsdat
}
\description{
Count mortality statistics in randomised controlled 
trials of treatments for chronic obstructive pulmonary disease (Woods et al. (2010), Table 1)
}
\keyword{datasets}
